refactor(docs-site): give the npm version url a module of its own
