+++ apps/docs-site/src/components/library-doc-page.tsx
@@ -16,7 +16,8 @@ import { buildGuidesHref } from '@/lib/guide-filters'
 import { getGuidesForPackage } from '@/lib/guides'
 import { markdownToHtml } from '@/lib/markdown'
 import { extractMermaidBlocks } from '@/lib/mermaid-utils'
-import { getPackageFacts, npmPackageUrl } from '@/lib/package-facts'
+import { npmPackageUrl } from '@/lib/npm-url'
+import { getPackageFacts } from '@/lib/package-facts'
 import { CAPABILITIES_SLOT, preparePackageReadme } from '@/lib/package-readme'
 import { readSectionLink } from '@/lib/readme-sections'
 import { buildRelatedReading } from '@/lib/related-reading'
